requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
scikit-learn>=1.3.0
librosa>=0.10.0
pydub>=0.25.0
numpy>=1.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6